Split models into class per file
