+++ src/annotation/models/__init__.py
@@ -1,0 +1,7 @@
+"""Defines the models of the application."""
+from .annotation import Annotation
+from .entry import Entry
+from .entrypage import EntryPage
+from .page import Page
+from .utils import extract_entry
+from .volume import Volume
